requirements.txt
moderngl 
glfw 
Pillow 
tinycss2 
numpy 
sounddevice 
soundfile